requirements.txt
# API & WEB Framework
fastapi
uvicorn[standard]
pydantic
python-multipart

# AI & LLM Integration
openai
langchain
python-dotenv

# Vector Database (Multimodal Search)
marqo

# Image Processing
rembg
Pillow
opencv-python-headless

# Data Handling
numpy
pandas

# Demo UI
streamlit

# added by health check
requests>=2.31.0
torch>=2.1.1
torchvision>=0.16.1
starlette>=0.27.0